requirements.txt
aiomysql==0.2.0
aioredis==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
cachetools==5.3.3
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-cache==0.1.0
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
idna==3.6
mysqlclient==2.2.4
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==1.7.1
PyMySQL==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
